refactor: move util viewmodel functions
