+++ app/src/main/kotlin/view/utils/CreateGraphDialogWindow.kt
@@ -16,6 +16,7 @@ import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.Dialog
 import androidx.compose.ui.window.DialogProperties
 import viewmodel.graph.SetupGraphViewModel
+import viewmodel.graph.createGraphFromTypesIndices
 
 @Composable
 fun CreateGraphDialogWindow(viewModel: SetupGraphViewModel) {
@@ -110,6 +111,7 @@ fun RadioColumn(
             radioOptions.forEachIndexed { index, option ->
                 Row(
                     verticalAlignment = Alignment.CenterVertically,
+                    horizontalArrangement = Arrangement.spacedBy(8.dp),
                     modifier = Modifier
                         .padding(vertical = 4.dp)
                         .fillMaxWidth()
@@ -122,7 +124,6 @@ fun RadioColumn(
                             selectedColor = MaterialTheme.colors.secondary
                         )
                     )
-                    Spacer(Modifier.width(8.dp))
                     Text(
                         text = option,
                         style = TextStyle(fontSize = 16.sp),
@@ -132,61 +133,4 @@ fun RadioColumn(
             }
         }
     }
-}
-
-
-@Composable
-fun createGraphFromTypesIndices(
-    viewModel: SetupGraphViewModel,
-    storedDataIndex: Int,
-    orientationIndex: Int,
-    weightnessIndex: Int
-) {
-    val storedData = when (storedDataIndex) {
-        0 -> SetupGraphViewModel.GraphType.Integer
-        1 -> SetupGraphViewModel.GraphType.UInteger
-        2 -> SetupGraphViewModel.GraphType.String
-        else -> SetupGraphViewModel.GraphType.Integer // default to integer
-    }
-
-    val graphStructure = when (orientationIndex) {
-        0 -> SetupGraphViewModel.GraphStructure.Undirected
-        1 -> SetupGraphViewModel.GraphStructure.Directed
-        else -> SetupGraphViewModel.GraphStructure.Undirected // default to undirected
-    }
-
-    val weight = when (weightnessIndex) {
-        0 -> SetupGraphViewModel.Weight.Unweighted
-        1 -> SetupGraphViewModel.Weight.Weighted
-        else -> SetupGraphViewModel.Weight.Unweighted // default to unweighted
-    }
-
-    return viewModel.createGraphAndApplyScreen(storedData, graphStructure, weight)
-}
-
-fun getGraphVMParameter(
-    storedDataType: Int,
-    structureType: Int,
-    weightType: Int
-): Triple<SetupGraphViewModel.GraphType, SetupGraphViewModel.GraphStructure, SetupGraphViewModel.Weight> {
-    val storedData = when (storedDataType) {
-        0 -> SetupGraphViewModel.GraphType.Integer
-        1 -> SetupGraphViewModel.GraphType.UInteger
-        2 -> SetupGraphViewModel.GraphType.String
-        else -> SetupGraphViewModel.GraphType.Integer // default to integer
-    }
-
-    val graphStructure = when (structureType) {
-        0 -> SetupGraphViewModel.GraphStructure.Undirected
-        1 -> SetupGraphViewModel.GraphStructure.Directed
-        else -> SetupGraphViewModel.GraphStructure.Undirected // default to directed
-    }
-
-    val weight = when (weightType) {
-        0 -> SetupGraphViewModel.Weight.Unweighted
-        1 -> SetupGraphViewModel.Weight.Weighted
-        else -> SetupGraphViewModel.Weight.Unweighted // default to weighted
-    }
-
-    return Triple(storedData, graphStructure, weight)
 }
